requirements.txt
requests
img2pdf
python-telegram-bot==13.15
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
